requirements.txt
pyasn1==0.4.8
selenium==3.141.0
pytz==2020.4
capmonster_python==1.3.2
fake_useragent==0.1.11
google_api_python_client==2.18.0
protobuf==3.17.3
